[Faucet] Adding 2 consumer queues for locking (#13512)
Separating out queues so that batch and normal transfers don't try to
use the same gas coin

## Description 

Currently there is a race condition where the 2 API calls `batch_gas`
and `gas` try to use the same coin. This is a PR to implement a
different queue so that they don't contest each other

## Test Plan 

How did you test the new or updated feature?

tested by deploying :) 
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes