[dapp-kit] add support for registering a burner wallet to the wallet provider (#13656)
## Description 

This PR adds support for the `enableUnsafeBurner` prop to the new
`WalletProvider` component. The base of the code is copied + pasted from
`wallet-kit` with some tweaks here and there. I abstracted the
functionality into a hook just to keep the WalletProvider component
clean.

## Test Plan 
- Automated tests
- Manual testing in example app that'll come later
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Randall-Mysten <109545725+randall-Mysten@users.noreply.github.com>