Stop depending on tombstones (#13609)
## Description 

The ParentSync trait has an explicit dependency to tombstones as it also
returns the tombstone as the result if it's the latest entry. Depending
on this trait will make it impossible to prune tombstones.
Currently only sui-execution v0 still depends on it, and no other places
should continue to depend on it.
This PR gets rid of such dependency except in sui-execution v0.
One thing I changed here which requires a close review is in the shared
object lock assignment path, where I believe we shouldn't need to use
parent sync.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes