[orchestrator] introduce the orchestractor crate (#13412)
## Description 

Original author of this work is @asonnino - I've mainly rewired
dependencies and removed any use of OpenSSL relying to pure Rust ssh
implementation in order to make it acceptable from our license checks +
some clean up.

The Orchestrator crate provides facilities for quickly deploying and
benchmarking the SUI codebase in a geo-distributed environment. Please
note that it is not intended for production deployments or as an
indicator of production engineering best practices. Its purpose is to
facilitate research projects by allowing benchmarking of (variants of)
the codebase and analyzing performance.

## Test Plan 

Did run against an AWS environment with multiple runs.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Alberto Sonnino <alberto@sonnino.com>