[dapp-kit] evolve wallet test mocks to support passing multiple accounts and do some re-organizing (#13733)
## Description 
I originally included these changes in the auto-connection PR, but that
PR was starting to get hard to review so I decided to pull this into its
own PR. This PR just moves some tests from `useWallet.test.tsx` to
`WalletProvider.test.tsx` since we're really testing the component logic
as opposed to the hook logic. It also modifies the `MockWallet` class to
support the creation of multiple accounts in tests which is needed for
the auto-connection logic PR.

(side note: I'm kind of re-working these mocks into the ideal state as I
go - re-factoring for passing in various feature sets will likely come
next 😄)

## Test Plan 
- Automated tests pass
- CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes