[refactor] consensus handler (#13705)
## Description 

Refactored consensus handler to pass as value to Narwhal since it's only
managed by the subscriber anyways and the certificate processing is
sequential. That also allow us to not use any Mutexes for internal state
management but rather act on a `&mut self` reference. Also did some
clean up so we do use the injected hostname map for metric reporting ,
but rather rely on the existing injected committee object to derive the
same info.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes