[EffectsV2 2/n] Record modified object ids only in execution results (#13783)
## Description 

In the current implementation, we track the input object metadata at
runtime, including object digest. Then in the end, we put that
information in objects_modified_at, which we could use to get the old
object information in effects v2.
There are two issues with this approach:
1. These information are somewhat redundant to what we already have and
record in mutable inputs and loaded child objects. Having another copy
is inefficient and leaves space for mistakes.
2. In fact, the object digest in the input object metadata is not
guaranteed to be the same as the original input object digest, since we
may be smashing gas prior to entering PT. So it's actually incorrect to
use that information.

This PR simplifies the modified_at in execution results to just
modified_objects as a list of IDs.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes