Move suivm to latest (#13814)
## Description 

Move suivm to latest. Suivm is getting in the way of a bunch of work
happening in the execution layer.
This is also showing the issues related to branches that are live for
long. It is at times inevitable but we need to be careful about what
people review and how to go back to latest.
Hopefully the work to make latest the working branch will make this less
of a problem. We'll see

## Test Plan 

Existing tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes